Merges 11 commits from main into pd, including:
- trained-roofline latency backend (#616)
- MaxOutputLen population + engine auto-fill (#621)
- MaxModelLen int64 + rope_scaling extraction (#606)
- MoE-aware roofline latency (#561)
- MaxModelLen enforcement + oracle knowledge boundary (#579, #587)
- Dead HardwareCalib fields removal (#596)
- Default example model switch to Qwen3-14B (#608)
- CI/CD updates (#600, #601, #607)

Conflict resolutions:
- CLAUDE.md: kept both INV-9 (main) and INV-PD-* (pd) invariants
- cmd/root.go: merged LengthCappedRequests counter + DroppedKVAllocations
- sim/bundle.go: added trained-roofline backend + kept disaggregation deciders
- sim/cluster/metrics.go: added LengthCappedRequests field
- docs: merged invariants and results documentation from both branches
- sim/cluster/disaggregation_test.go: added MaxModelLen param to NewModelHardwareConfig calls
- sim/cluster/cluster_event.go: rewrote comment to avoid INV-9 test false positive

+++ sim/internal/testutil/golden.go
@@ -38,7 +38,7 @@ type GoldenTestCase struct {
 	Seed                      int64         `json:"seed"`
 	MaxNumRunningReqs         int64         `json:"max-num-running-reqs"`
 	MaxNumScheduledTokens     int64         `json:"max-num-scheduled-tokens"`
-	MaxModelLen               int           `json:"max-model-len"`
+	MaxModelLen               int64         `json:"max-model-len"`
 	TotalKVBlocks             int64         `json:"total-kv-blocks"`
 	BlockSizeInTokens         int64         `json:"block-size-in-tokens"`
 	LongPrefillTokenThreshold int64         `json:"long-prefill-token-threshold"`
